Improved device table. Fixed video display ratio.
Now using a datatable to show the list of devices rather than the previous very simply table. Piped through various extra fields so that they can be optionally shown as well.  Piped through display of the iPhone model name to the table. ( requires the latest updated to ios_remote_provider )

Fixed the updateDeviceInfo function, so that changes to device info now properly flow to the database.  Added additional information output to the more detailed device info page.

Corrected the video display to use the proper ratio from the device itself.

Added username auto-select on load of the login page, to make it easier/quicker to login.