Improve key entry
Make key entry wait to return. Previously it was returning immediately and that was causing keys to be entered out of order on device.

Improved how keyCode was/is being fetched. It wasn’t cross-browser consistent and was using deprecated event.which.

Move more things ( everything? ) to use ujsonin v2 module. There was a mix which should not be.