Simplify UI
Added "themes" to the UI, and set the default theme to a newly created "simple" theme that has no side menu.

In the new simple theme, cleanedup the table display some:
1. Changed the column selection text to just be "Columns" and removed select/deselect of the first column.
2. Changed the first column to better display the status of the device and whether it is in use by a user or not.
3. Removed Search: from being before the search box and placed the text inside the box to be cleaner.
4. Added display of iOS version by default.

Changed the default action on clicking on a row to go directly to device video instead of going to the device info page. As a result the device info page has no way to reach it currently in the simple theme.

There is also no way to access logout currently in the simple theme. Need to add it back in. TODO

Tweaked the video display page to display some limited device info to the right of the device video.
Added a back arrow on the page to go back to the device list since there is no longer a menu in the simple theme to do so.
Added a spinner beneath the video display to indicate to the user that video is loading for the short bit before video begins streaming. Need to also provide detection in the backend when the video app is down and display a message and/or auto restart the video app. That is not yet done, so it is currently possible to get a spinning wait icon for video and no video to ever show. TODO.