Swagger. Video updates. API updates. Inspect page.
Added swagger for API documentation and ease of viewing of it.

Added notes displayed next to video stream for displaying important notes to users on how to use the system.

Reduced vertical resolution of video to 850 to almost match the output of the updated video app.

Changed the API to clean it up a bunch.

Added inspect page and related source command. This displays a wireframe version of device UI.

Improved performance of video display some. Removed/hid the loading icon behind the video for the moment as it
was causing constant re-evaluations of css and high cpu usage.

Updated style of "device is reserved" page for the simple theme.