Video stream tweaks
Updated video stream connection to client a bit to split receiving of frames from provider from sending of frames to client. This change by itself should make little to no difference but is in preperation for timing measurements of frame sending to skip sending frames if they are not being received by the client fast enough.