Use a channel instead of lock for frame sending
Changed adaptive frame sending logic to use a channel instead of a lock for receiving frames and then sending them back out at the determined bandwidth.

This new logic is more straightforward and less buggy than the previous method. It is also easier to understand what it is doing exactly now.